Lot of refactor

+++ app/src/main/java/com/example/tasky/agenda/agenda_data/notification/NotificationService.kt
@@ -6,11 +6,12 @@ import androidx.work.WorkManager
 import androidx.work.workDataOf
 import com.example.tasky.agenda.agenda_domain.NotificationScheduler
 import com.example.tasky.agenda.agenda_domain.model.AgendaItem
+import com.example.tasky.core.presentation.DateUtils.toLong
 import java.util.concurrent.TimeUnit
 
 class NotificationService(private val context: Context): NotificationScheduler {
     override fun schedule(agendaItem: AgendaItem) {
-        val delay = agendaItem.remindAt - System.currentTimeMillis()
+        val delay = (agendaItem.remindAt.toLong() - System.currentTimeMillis())
 //        Timber.d("DDD - NotificationService: ${agendaItem.remindAt.toLocalDateTime()} | ${System.currentTimeMillis().toLocalDateTime()}")
         val workRequest = OneTimeWorkRequestBuilder<NotificationWorker>()
             .setInputData(workDataOf("title" to agendaItem.title, "message" to agendaItem.description))
